refactor: Refactor test files
